South Carolina Food Deserts

The dataset of census tracts deemed to be food deserts by the USDA

fooddeserts_processed

Format

A data frame with 1091 rows and 149 columns

census_tract

Census tract

state

State

county

County

urban

Urban tract

pop2010

Population, tract total

ohu2010

Housing units, total

group_quarters_flag

Group quarters, tract with high share

numgqtrs

Group quarters, tract population residing in, number

pctgqtrs

Group quarters, tract population residing in, share

lila_tracts_1and10

Low income and low access tract measured at 1 mile for urban areas and 10 miles for rural areas

lila_tracts_halfand10

Low income and low access tract measured at 1/2 mile for urban areas and 10 miles for rural areas

lila_tracts_1and20

Low income and low access tract measured at 1 mile for urban areas and 20 miles for rural areas

lila_tracts_vehicle

Low income and low access tract using vehicle access or low income and low access tract measured at 20 miles

hunv_flag

Vehicle access, tract with low vehicle access

low_income_tracts

Low income tract

poverty_rate

Tract poverty rate

median_family_income

Tract median family income

la1and10

Low access tract at 1 mile for urban areas and 10 miles for rural areas

l_ahalfand10

Low access tract at 1/2 mile for urban areas and 10 miles for rural areas

la1and20

Low access tract at 1 mile for urban areas and 20 miles for rural areas

la_tracts_half

Low access tract at 1/2 mile

la_tracts1

Low access tract at 1 mile

la_tracts10

Low access tract at 10 miles

la_tracts20

Low access tract at 20 miles

la_tracts_vehicle_20

Low access tract using vehicle access and at 20 miles in rural areas

lapop1_10

Low access, population at 1 mile for urban areas and 10 miles for rural areas, number

lapop05_10

Low access, population at 1/2 mile for urban areas and 10 miles for rural areas, number

lapop1_20

Low access, population at 1 mile for urban areas and 20 miles for rural areas, number

lalowi1_10

Low access, low-income population at 1 mile for urban areas and 10 miles for rural areas, number

lalowi05_10

Low access, low-income population at 1/2 mile for urban areas and 10 miles for rural areas, number

lalowi1_20

Low access, low-income population at 1 mile for urban areas and 20 miles for rural areas, number

lapophalf

Low access, population at 1/2 mile, number

lapophalfshare

Low access, population at 1/2 mile, share

lalowihalf

Low access, low-income population at 1/2 mile, number

lalowihalfshare

Low access, low-income population at 1/2 mile, share

lakidshalf

Low access, children age 0-17 at 1/2 mile, number

lakidshalfshare

Low access, children age 0-17 at 1/2 mile, share

laseniorshalf

Low access, seniors age 65+ at 1/2 mile, number

laseniorshalfshare

Low access, seniors age 65+ at 1/2 mile, share

lawhitehalf

Low access, White population at 1/2 mile, number

lawhitehalfshare

Low access, White population at 1/2 mile, share

lablackhalf

Low access, Black or African American population at 1/2 mile, number

lablackhalfshare

Low access, Black or African American population at 1/2 mile, share

laasianhalf

Low access, Asian population at 1/2 mile, number

laasianhalfshare

Low access, Asian population at 1/2 mile, share

lanhopihalf

Low access, Native Hawaiian or Other Pacific Islander population at 1/2 mile, number

lanhopihalfshare

Low access, Native Hawaiian or Other Pacific Islander population at 1/2 mile, share

laaianhalf

Low access, American Indian or Alaska Native population at 1/2 mile, number

laaianhalfshare

Low access, American Indian or Alaska Native population at 1/2 mile, share

laomultirhalf

Low access, Other/Multiple race population at 1/2 mile, number

laomultirhalfshare

Low access, Other/Multiple race population at 1/2 mile, share

lahisphalf

Low access, Hispanic or Latino population at 1/2 mile, number

lahisphalfshare

Low access, Hispanic or Latino population at 1/2 mile, share

lahunvhalf

Vehicle access, housing units without and low access at 1/2 mile, number

lahunvhalfshare

Vehicle access, housing units without and low access at 1/2 mile, share

lasnaphalf

Low access, housing units receiving SNAP benefits at 1/2 mile, number

lasnaphalfshare

Low access, housing units receiving SNAP benefits at 1/2 mile, share

lapop1

Low access, population at 1 mile, number

lapop1share

Low access, population at 1 mile, share

lalowi1

Low access, low-income population at 1 mile, number

lalowi1share

Low access, low-income population at 1 mile, share

lakids1

Low access, children age 0-17 at 1 mile, number

lakids1share

Low access, children age 0-17 at 1 mile, share

laseniors1

Low access, seniors age 65+ at 1 mile, number

laseniors1share

Low access, seniors age 65+ at 1 mile, share

lawhite1

Low access, White population at 1 mile, number

lawhite1share

Low access, White population at 1 mile, share

lablack1

Low access, Black or African American population at 1 mile, number

lablack1share

Low access, Black or African American population at 1 mile, share

laasian1

Low access, Asian population at 1 mile, number

laasian1share

Low access, Asian population at 1 mile, share

lanhopi1

Low access, Native Hawaiian and Other Pacific Islander population at 1 mile, number

lanhopi1share

Low access, Native Hawaiian and Other Pacific Islander population at 1 mile, share

laaian1

Low access, American Indian and Alaska Native population at 1 mile, number

laaian1share

Low access, American Indian and Alaska Native population at 1 mile, share

laomultir1

Low access, Other/Multiple race population at 1 mile, number

laomultir1share

Low access, Other/Multiple race population at 1 mile, share

lahisp1

Low access, Hispanic or Latino population at 1 mile, number

lahisp1share

Low access, Hispanic or Latino population at 1 mile, share

lahunv1

Vehicle access, housing units without and low access at 1 mile, number

lahunv1share

Vehicle access, housing units without and low access at 1 mile, share

lasnap1

Low access, housing units receiving SNAP benefits at 1 mile, number

lasnap1share

Low access, housing units receiving SNAP benefits at 1 mile, share

lapop10

Low access, population at 10 miles, number

lapop10share

Low access, population at 10 miles, share

lalowi10

Low access, low-income population at 10 miles, number

lalowi10share

Low access, low-income population at 10 miles, share

lakids10

Low access, children age 0-17 at 10 miles, number

lakids10share

Low access, children age 0-17 at 10 miles, share

laseniors10

Low access, seniors age 65+ at 10 miles, number

laseniors10share

Low access, seniors age 65+ at 10 miles, share

lawhite10

Low access, White population at 10 miles, number

lawhite10share

Low access, White population at 10 miles, share

lablack10

Low access, Black or African American population at 10 miles, number

lablack10share

Low access, Black or African American population at 10 miles, share

laasian10

Low access, Asian population at 10 miles, number

laasian10share

Low access, Asian population at 10 miles, share

lanhopi10

Low access, Native Hawaiian and Other Pacific Islander population at 10 miles, number

lanhopi10share

Low access, Native Hawaiian and Other Pacific Islander population at 10 miles, share

laaian10

Low access, American Indian and Alaska Native population at 10 miles, number

laaian10share

Low access, American Indian and Alaska Native population at 10 miles, share

laomultir10

Low access, Other/Multiple race population at 10 miles, number

laomultir10share

Low access, Other/Multiple race population at 10 miles, share

lahisp10

Low access, Hispanic or Latino population at 10 miles, number

lahisp10share

Low access, Hispanic or Latino population at 10 miles, share

lahunv10

Vehicle access, housing units without and low access at 10 miles, number

lahunv10share

Vehicle access, housing units without and low access at 10 miles, share

lasnap10

Low access, housing units receiving SNAP benefits at 10 miles, number

lasnap10share

Low access,housing units receiving SNAP benefits at 10 miles, share

lapop20

Low access, population at 20 miles, number

lapop20share

Low access, population at 20 miles, share

lalowi20

Low access, low-income population at 20 miles, number

lalowi20share

Low access, low-income population at 20 miles, share

lakids20

Low access, children age 0-17 at 20 miles, number

lakids20share

Low access, children age 0-17 at 20 miles, share

laseniors20

Low access, seniors age 65+ at 20 miles, number

laseniors20share

Low access, seniors age 65+ at 20 miles, share

lawhite20

Low access, White population at 20 miles, number

lawhite20share

Low access, White population at 20 miles, share

lablack20

Low access, Black or African American population at 20 miles, number

lablack20share

Low access, Black or African American population at 20 miles, share

laasian20

Low access, Asian population at 20 miles, number

laasian20share

Low access, Asian population at 20 miles, share

lanhopi20

Low access, Native Hawaiian and Other Pacific Islander population at 20 miles, number

lanhopi20share

Low access, Native Hawaiian and Other Pacific Islander population at 20 miles, share

laaian20

Low access, American Indian and Alaska Native population at 20 miles, number

laaian20share

Low access, American Indian and Alaska Native population at 20 miles, share

laomultir20

Low access, Other/Multiple race population at 20 miles, number

laomultir20share

Low access, Other/Multiple race population at 20 miles, share

lahisp20

Low access, Hispanic or Latino population at 20 miles, number

lahisp20share

Low access, Hispanic or Latino population at 20 miles, share

lahunv20

Vehicle access, housing units without and low access at 20 miles, number

lahunv20share

Vehicle access, housing units without and low access at 20 miles, share

lasnap20

Low access, housing units receiving SNAP benefits at 20 miles, number

lasnap20share

Low access, housing units receiving SNAP benefits at 20 miles, share

tractlowi

Tract low-income population, number

tractkids

Tract children age 0-17, number

tractseniors

Tract seniors age 65+, number

tractwhite

Tract White population, number

tractblack

Tract Black or African American population, number

tractasian

Tract Asian population, number

tractnhopi

Tract Native Hawaiian and Other Pacific Islander population, number

tractaian

Tract American Indian and Alaska Native population, number

tractomultir

Tract Other/Multiple race population, number

tracthispanic

Tract Hispanic or Latino population, number

tracthunv

Tract housing units without a vehicle, number

tractsnap

Tract housing units receiving SNAP benefits, number

tract_latitude

Tract Centroid Latitude

tract_longitude

Tract Centroid Longitude

Source

https://www.ers.usda.gov/data-products/food-access-research-atlas/download-the-data/