South Carolina Farmers and Roadside Markets

The dataset of the farmers and roadside markets in South Carolina #' NEED TO CHECK for any requirements in redistributing data such as a notice copyright, etc

markets_processed

Format

A data frame with 337 rows and 38 columns

x

X Coordinate

y

Y Coordinate

objectid

Object ID

name

The Name of the Market

street

Street Address

city

City

state

State

zip

ZIP Code

county

County

mail_phone

Phone Number

website

Website

open_jan

Open in January

open_feb

Open in February

open_mar

Open in March

open_apr

Open in April

open_may

Open in May

open_jun

Open in June

open_jul

Open in July

open_aug

Open in August

open_sep

Open in September

open_oct

Open in October

open_nov

Open in November

open_dec

Open in December

cash

Accepts Cash Payment

check

Accepts Check Payment

credit

Accepts Credit Payment

debit

Accepts Debit Payment

other_pay

Accepts Other Forms of Payment

wic

Accepts WIC Payments

senior

Accepts Senior Vouchers

snap_ebt

Accepts SNAP/EBT Payment

snap_trans

SNAP Transaction Type

h_bucks

Accepts Healthy Bucks Payment

geoid

GEOID

tract_latitude

Tract Centroid Latitude

tract_longitude

Tract Centroid Longitude

lat

Latitude

lon

Longitude

Source

https://sc-department-of-health-and-environmental-control-gis-sc-dhec.hub.arcgis.com/datasets/cdb563626aaf4477b7ea591fec10a3e2_0/explore