South Carolina Pharmacy Network

The dataset of the pharmacies in South Carolina #' NEED TO CHECK for any requirements in redistributing data such as a notice copyright, etc

pharmacies_processed

Format

A data frame with 954 rows and 54 columns

npi

National Provider Identifier Number

basic_organization_name

The Name of the Pharmacy Organization

npi_2

National Provider Identifier Number

hours_of_operation

Hours of Operation

medicaid_id

Medicaid Identifier

spoken_languages

Spoken Languages

e_rx

Offers e-Rx

delivery

Offers Delivery

compound

Offers Compounds

drive_thru

Offers Drive-Thru

vaccines

Offers Vaccines

handicap_access

Offers Handicap Access

cultural_comp_training

Provides Cultural Competence Training

accepting_new_customers

Accepting New Customers

website

Website

basic_organizational_subpart

Organizational Subpart?

basic_enumeration_date

Enumeration Date

basic_last_updated

Date Last Updated

basic_status

Status (Active)

basic_authorized_official_credential

Authorized Official Credential

basic_authorized_official_first_name

Authorized Official First Name

basic_authorized_official_last_name

Authorized Official Last Name

basic_authorized_official_telephone_number

Authorized OFficial Telephone Number

basic_authorized_official_title_or_position

Authorized OFficial Title/Position

basic_name

Pharmacy Name

basic_parent_organization_ein

Parent Organization EIN

basic_parent_organization_legal_business_name

Parent Organization

basic_authorized_official_middle_name

Authorized Official Middle Name

basic_authorized_official_name_prefix

Authorized Official Name Prefix

basic_authorized_official_name_suffix

Authorized Official Name Suffix

basic_certification_date

Certification Date

basic_name_prefix

Other Official Name Prefix

basic_first_name

Other Official First Name

basic_last_name

Other Official Last Name

basic_middle_name

Other Official Middle Name

basic_credential

Other Official redentials

basic_sole_proprieter

Sole Proprietor

basic_gender

Other Official Gender

addresses_country_code

Country Code

addresses_country_name

Country

addresses_address_purpose

Address Purpose

addresses_address_type

Address Type

addresses_address_1

Street Address

addresses_address_2

Street Address 2

addresses_city

City

addresses_state

State

addresses_postal_code

ZIP Code

addresses_telephone_number

Telephone Number

addresses_fax_number

Fax Number

geoid

GEOID

tract_latitude

Tract Centroid Latitude

tract_longitude

Tract Centroid Longitude

lat

Latitude

lon

Longitude

Source

https://www.selecthealthofsc.com/pdf/provider/pharmacy/pharmacy-network.pdf